animalrescue/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
black==24.8.0
click==8.1.7
decorator==5.1.1
Django==5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.0.1
ipdb==0.13.13
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
